Don't fail entire upload process if just one file does not exist

+++ mentor_upload_worker/src/mentor_upload_process/process.py
@@ -211,20 +211,25 @@ def process_answer_video(
             s3 = _create_s3_client()
             s3_bucket = _require_env("STATIC_AWS_S3_BUCKET")
             for media_type, tag, file_name, content_type, file in media_uploads:
-                item_path = f"{video_path_base}{file_name}"
-                media.append(
-                    {
-                        "type": media_type,
-                        "tag": tag,
-                        "url": item_path,
-                    }
-                )
-                s3.upload_file(
-                    str(file),
-                    s3_bucket,
-                    item_path,
-                    ExtraArgs={"ContentType": content_type},
-                )
+                if path.isfile(file):
+                    item_path = f"{video_path_base}{file_name}"
+                    media.append(
+                        {
+                            "type": media_type,
+                            "tag": tag,
+                            "url": item_path,
+                        }
+                    )
+                    s3.upload_file(
+                        str(file),
+                        s3_bucket,
+                        item_path,
+                        ExtraArgs={"ContentType": content_type},
+                    )
+                else:
+                    import logging
+
+                    logging.error(f"Failed to find file at {file}")
             update_status(
                 StatusUpdateRequest(
                     mentor=mentor,
@@ -302,58 +307,59 @@ def process_transfer_video(req: ProcessTransferRequest, task_id: str):
         )
     )
     for m in media:
-        if not m.get("needsTransfer", False):
-            continue
-        typ = m.get("type", "")
-        tag = m.get("tag", "")
-        root_ext = "vtt" if typ == "subtitles" else "mp4"
-        file_path, headers = urllib.request.urlretrieve(m.get("url", ""))
-        try:
-            item_path = f"videos/{mentor}/{question}/{datetime.utcnow().strftime('%Y%m%dT%H%M%SZ')}/{tag}.{root_ext}"
-            s3 = _create_s3_client()
-            s3_bucket = _require_env("STATIC_AWS_S3_BUCKET")
-            content_type = "text/vtt" if typ == "subtitles" else "video/mp4"
-            s3.upload_file(
-                file_path,
-                s3_bucket,
-                item_path,
-                ExtraArgs={"ContentType": content_type},
-            )
-            m["needsTransfer"] = False
-            m["url"] = item_path
-            update_status(
-                StatusUpdateRequest(
-                    mentor=mentor,
-                    question=question,
-                    task_id=task_id,
-                    status="TRANSFER_IN_PROGRESS",
-                    transcript=transcript,
-                    media=media,
-                )
-            )
-            update_media(MediaUpdateRequest(mentor=mentor, question=question, media=m))
-        except Exception as x:
-            import logging
-
-            logging.exception(x)
-            update_status(
-                StatusUpdateRequest(
-                    mentor=mentor,
-                    question=question,
-                    task_id=task_id,
-                    status="TRANSFER_FAILED",
-                    transcript=transcript,
-                    media=media,
-                )
-            )
-        finally:
+        if m.get("needsTransfer", False):
+            typ = m.get("type", "")
+            tag = m.get("tag", "")
+            root_ext = "vtt" if typ == "subtitles" else "mp4"
+            file_path, headers = urllib.request.urlretrieve(m.get("url", ""))
             try:
-                remove(file_path)
+                item_path = f"videos/{mentor}/{question}/{datetime.utcnow().strftime('%Y%m%dT%H%M%SZ')}/{tag}.{root_ext}"
+                s3 = _create_s3_client()
+                s3_bucket = _require_env("STATIC_AWS_S3_BUCKET")
+                content_type = "text/vtt" if typ == "subtitles" else "video/mp4"
+                s3.upload_file(
+                    file_path,
+                    s3_bucket,
+                    item_path,
+                    ExtraArgs={"ContentType": content_type},
+                )
+                m["needsTransfer"] = False
+                m["url"] = item_path
+                update_status(
+                    StatusUpdateRequest(
+                        mentor=mentor,
+                        question=question,
+                        task_id=task_id,
+                        status="TRANSFER_IN_PROGRESS",
+                        transcript=transcript,
+                        media=media,
+                    )
+                )
+                update_media(
+                    MediaUpdateRequest(mentor=mentor, question=question, media=m)
+                )
             except Exception as x:
                 import logging
 
-                logging.error(f"failed to delete file '{file_path}'")
                 logging.exception(x)
+                update_status(
+                    StatusUpdateRequest(
+                        mentor=mentor,
+                        question=question,
+                        task_id=task_id,
+                        status="TRANSFER_FAILED",
+                        transcript=transcript,
+                        media=media,
+                    )
+                )
+            finally:
+                try:
+                    remove(file_path)
+                except Exception as x:
+                    import logging
+
+                    logging.error(f"failed to delete file '{file_path}'")
+                    logging.exception(x)
     update_status(
         StatusUpdateRequest(
             mentor=mentor,
